web 완전판

+++ Reservation_Web/CarReservationWeb/Models/AccountParking.cs
@@ -16,7 +16,7 @@ public partial class AccountParking
 	[DataType(DataType.Password)]
 	public string Password { get; set; } = null!;
 
-	public sbyte Admin { get; set; }
+	public int Admin { get; set; }
 
 	public string? NfcRegistered { get; set; }
 }
